+++ doc/CHANGES.md
@@ -3,6 +3,7 @@
 
 ## Highlighted Features
 
+* Remove rendering of exceptions in admin interface (Gareth Rees)
 * Pass through sign-in form if a user is already signed in (Gareth Rees)
 * Make the event history table responsive (Miroslav Schlossberg)
 * Fix bug that prevented private requests from being published across the whole